**OpenWeatherApiClient**
- Added APICounter bean to class

- Moved checkDateAndAddCounter method to APICounter class

 - Moved checkAPILimitIsNotReached method to APICounter class

 - Removed if-statement from getOpenWeatherAPIResponse

 - Moved OpenFileAndUpdateCounter method to APICounter class

 **APICounter**

 - Added APICounter class

 - Added methods mentioned above

 - Added new logic to CheckIfAPILimitIsReached to now compare date in counter.txt and currentDate

 **ConditionResponseClient**

 - Added APICounter bean

 - Added logic in getConditionResponseFromDateTime to break out of loop if API limit is reached
